refactor: layout of ArticleList
